requirements.txt
numpy 
tensorflow
Pillow 
opencv-python
notebook
jupyter 
accelerate
torch
transformers
huggingface_hub

